requirements.txt
torch
boto3
torchvision
Pillow
scikit-learn
tqdm
optuna
nni